Add module content.
At the moment the module can read an image file using Imager, and get
the bytes of the image as greyscale, and it can calculate the values
of the image pixels using a simplified version of the algorithm used
in libpuzzle. It can also write out a PNG image of the data being
analysed for testing purposes.